System/nanoTime instead of System/currentTimeMillis
The precision is slightly improved, but it comes with a tradeoff that I'm not sure is worth it.

`System/nanoTime` is more accurate than `currentTimeMillis`, but it is definitely more expensive. For some instances, I can see the added precision being valuable (if you wanted to detect really small amounts of jitter in a datadog monitor for instance), but `currentTimeMillis` might be the better default option.